Fix error in example_1
This test case had an error:
test_FindFunction_WhichIsBroken_ShouldReturnZeroIfItemIsNotInList_WhichWorksEvenInOurBrokenCode
It was supposed to be a list of values that are NOT in the list,
and none of them should be found.  It incorrectly included '1'
which is a value in the list.

The compile option -Wno-misleading-indentation was also added to
remove a compiler warning produced by gcc 7.3.0